# (#5118) - Revert #4867 to fix performance regression