@nolanlawson 2016-05-03
Backstory: https://github.com/pouchdb/pouchdb/issues/5118