@daleharvey 2016-05-04
So this code always worried me a little, premature caching can often get in the way of other optimizations, it means doing more work. Given the regression for the more common case 👍 on taking it out